Moved svn:externals parse test into a test.
This utilizing unittest and is self-contained within the exectuable.
For now, there is a global variable that controls whether the tests or
run or the program is run.  This should be fixed at some point.

Note that it would nice for the tests to be separate, but the naming
convention of Git subcommands is such that it seems like it may be
impossible.